Implemented more CIFS mount options under Linux as well as a whitelist of mount.cifs options that are save to be entered via the "Additional options" line edit in the configuration dialog.
The new (and removed) mount options are:
- 'cache=none|strict|loose' replacing 'directio' that is deprecated as of
  kernel 3.7 and was thus removed. In the configuration dialog the respective
  setting is named 'Cache mode'.
- 'vers=1.0|2.0|2.1|3.0'. In the configuration dialog the respective setting
  is named 'SMB protocol version'.
- 'forceuid'. In the configuration dialog the respective setting is named
  'Definitely assign the UID'.
- 'forcegid'. In the configuration dialog the respective setting is named
  'Definitely assign the GID'.

To ensure a seamless transition, configuration updates are provided.

To enforce proper checking of the additional options provided by the user,
the "Additional options" line edit has been set read-only and the user must
use the new edit button to be able to enter the options. All options that
are not whitelisted - i.e. they have already been defined elsewhere or they
pose a potential security risk - will be removed from the list upon commit.
This is, among others, the proper fix to the security issue reported
by Heiner Markert (aka CVE-2014-2581) and initially fixed in version 1.1.2.