Improve the handling of online and offline states.
In the case the network was down when the application was started,
nothing happened when the system went online. This is fixed now.

The initial update of the network configuration can be done by the
Smb4KHardwareInterface class. There is no need for the updateNetworkConfig()
function anymore.

When the network session is set up, emit the new networkSessionInitialized()
signal (instead of the old networkConfigUpdated() signal). You can use this
to trigger initial actions that are meant to be performed only when the system
is online.