Fixed several issues and crashes.
The crashes that occurred when remounting shares has been eliminated.

The notification of the user after a mount or unmount has been fixed.

The unmounting of shares when the system went offline has been fixed. For this to work,
you also need the improved mount helper.

Fixed hanging of the application in several places.

BUG: 415165
BUG: 419658
BUG: 415726