Simplify the installation of translations
Both ki18n_install and kdoctools_install are safe to use
if the parameter does not exist (they would be noop in that case),
so remove the guard.
Also, remove the additional guard about KDocTools being found
(which would have needed an update to KF6) because KDocTools
is listed as mandatory requirement anyway.