The login credentials are now stored using QtKeychain. This allows the usage of other secure storage solutions apart from KWallet.
Since QtKeychain does not provide access to all login credentials Smb4K stored, we lost the
capability to edit the login credentials in the configuration dialog. You need to use the
respective editor program (i.e. kwalletmanager, KeePassXC, ...) to edit or remove the login
credentials.

The default login now has become profile-dependent. It can be defined directly in the
configuration dialog and is automatically changed when the user changed the active profile.

BUG: 486474